requirements.txt
# Minimal dependencies for Vercel deployment
fastapi==0.104.1
pydantic==2.4.2
python-multipart==0.0.6
requests==2.31.0

# Image processing for pavement analysis
Pillow==10.1.0
numpy==1.24.3

# Azure Digital Twins (lightweight)
azure-digitaltwins-core==1.2.0
azure-identity==1.15.0